Move the PlantUML exporter into its own namespace.
Mostly because we're shortly going to have a Doctrine connector too...
